feat: Add climate sensor integration with configuration and data handling

+++ src/JaamWeb.h
@@ -32,7 +32,7 @@ private:
     String getMeta();
     String getStyles();
     String getScripts();
-    String getParameterHtml(const char* name, int min, int max, int step, int value, const char* label);
+    String getParameterHtml(const char* name, float min, float max, float step, float value, const char* label);
     String getBoolParameterHtml(const char* name, bool value, const char* label);
     String getColorPickerHtml(const char* name, const char* value, const char* label);
     String getDropdownHtml(const String& name, const String& label, Type settingKey, SettingListItem items[], int itemCount);
